Finishing step 1 of the our Plug investigation
Unfortunately, we have arrived at a cliffhanger at the end of this first
step because we still haven't discovered what is going on with Plug.
Other than saying that we are created a function call to Cowboy, it is
unclear what else we have gained with Plug, if anything. Hopefully, this
will become more evident as with move forward.